requirements.txt
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0
django-filter==2.2.0
djangorestframework==3.10.3
importlib-metadata==1.3.0
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
more-itertools==8.0.2
packaging==19.2
pkg-resources==0.0.0
pluggy==0.13.1
py==1.8.0
pycodestyle==2.5.0
pylint==2.4.4
pyparsing==2.4.5
pytest==5.3.1
python-decouple==3.3
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
static3==0.7.0
typed-ast==1.4.0
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.6.0
